fix(audit): repair what the remediation broke
A confirmation pass over the fixes themselves, which is where the damage was.

THE RESET GUARD WAS VACUOUS. `currentShaderParams` is an alias onto the
editor's `shaderParams`, and the editor assigns the default map to it BEFORE
emitting — so comparing `defaults` against it compared an object with itself and
was true every time. The guard collapsed to "the event owns no parameters",
making Reset a silent no-op on every inheriting event, including the one case
the comment beside it says must still write: an ancestor whose parameters are
tuned away from the pack defaults. It now reads the resolved map from the tree,
before the staging.

The curve refresh added for a refused write passed no `selfDriven`, so the card
treated its own refusal as an outside edit: it cleared the very latch that had
just refused the write, and could take the close-the-editor branch and fold the
timing section away under the user who had just picked a curve.

THE WARN ARM BECAME UNREACHABLE. Making the cascade walk agree with the resolver
meant it can never disagree with the resolver, so inside a branch gated on the
resolve being empty the "covered" verdict is false by construction. Rather than
leave a dead branch or reintroduce a predicate that contradicts resolution, the
walk now answers a question that is still worth asking — WHICH entry explains
the empty resolve — and both arms are debug, because neither an explicit None
nor an absent override is a fault.

The value bounds ran BEFORE the field allowlist, so junk keys consumed the entry
cap; since a QVariantMap iterates sorted, 256 keys sorting before "curve" would
have dropped the user's real edit. Allowlisting first caps the map at six by
construction.

Two files crossed the size ceiling on doc comments the remediation added. The
controller header's design essays move to the .cpp beside the code they
describe, and the group-write tests split along a real seam: everything that is
a refusal, a bound, or a failure is now test_animations_group_write_bounds.cpp,
with its own target above the isolation marker.

Corrected claims, each verified against the code rather than reworded: the retry
comment had its mechanism backwards (the stale reply is the NEWEST generation
and IS processed, which is harmless because the refresh is wholesale); the
baseline note said no writer emits a baseline when toJson always does, the true
statement being that nothing calls setBaseline; clearOverride's doc still said
it emits only on a successful removal; the allowlist was described as six keys
plus two; the lifetime note contradicted itself about a rule that does hold; and
two signal docs disagreed with each other about what a settings reload emits.
